+++ client/.idea/misc.xml
@@ -47,6 +47,7 @@
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_detail_plan.xml" value="0.20104166666666667" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_home_plan.xml" value="0.33" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_map_search_recycler.xml" value="0.2786458333333333" />
+        <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_participant.xml" value="0.39114583333333336" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_paticipant.xml" value="0.5" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_plan.xml" value="0.3046875" />
         <entry key="..\:/Users/jaino/2022_2_WAP_APP_TEAM1/client/app/src/main/res/layout/item_plan_flight.xml" value="0.22" />
