+++ client/app/src/main/java/com/example/witt/domain/model/use_case/plan/PlanStateModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.plan
+package com.example.witt.domain.model.use_case.plan
 
 
 data class PlanStateModel (
