+++ client/app/src/main/java/com/example/witt/utils/DefaultTemplate.kt
@@ -1,6 +1,6 @@
 package com.example.witt.utils
 
-import com.example.witt.domain.model.plan.PlanStateModel
+import com.example.witt.domain.model.use_case.plan.PlanStateModel
 import com.kakao.sdk.template.model.*
 
 object DefaultTemplate {
