+++ client/app/src/main/res/layout/item_home_plan.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="item"
-            type="com.example.witt.domain.model.plan.PlanStateModel" />
+            type="com.example.witt.domain.model.use_case.plan.PlanStateModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
