+++ client/app/src/main/res/layout/item_plan.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="item"
-            type="com.example.witt.domain.model.plan.adapter.PlanModel" />
+            type="com.example.witt.domain.model.use_case.plan.PlanModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
